Changes to annotation layer
