+++ client/app/reader/Pdf.jsx
@@ -99,7 +99,8 @@ export class Pdf extends React.PureComponent {
 const mapStateToProps = (state, props) => {
   return {
     ...state.readerReducer.ui.pdf,
-    ..._.pick(state.annotationsLayer, 'placingAnnotationIconPageCoords'),
+    ...state.annotationLayer.ui.pdf,
+    ..._.pick(state.annotationLayer, 'placingAnnotationIconPageCoords'),
     rotation: _.get(state.readerReducer.documents, [props.documentId, 'rotation']),
     sidebarHidden: state.readerReducer.ui.pdf.hidePdfSidebar
   };
