+++ client/app/reader/PdfPage.jsx
@@ -328,7 +328,7 @@ const mapDispatchToProps = (dispatch) => ({
 const mapStateToProps = (state, props) => {
   return {
     pageDimensions: _.get(state.readerReducer.pageDimensions, [`${props.file}-${props.pageIndex}`]),
-    isPlacingAnnotation: state.readerReducer.ui.pdf.isPlacingAnnotation,
+    isPlacingAnnotation: state.annotationLayer.ui.pdf.isPlacingAnnotation,
     rotation: _.get(state.readerReducer.documents, [props.documentId, 'rotation'], 0),
     searchText: searchText(state, props),
     currentMatchIndex: getCurrentMatchIndex(state, props),
