+++ client/app/reader/PdfSidebar.jsx
@@ -255,7 +255,7 @@ PdfSidebar.propTypes = {
 const mapStateToProps = (state, ownProps) => {
   return {
     ..._.pick(state.readerReducer.ui, 'placedButUnsavedAnnotation', 'selectedAnnotationId'),
-    comments: makeGetAnnotationsByDocumentId(state.readerReducer)(ownProps.doc.id),
+    comments: makeGetAnnotationsByDocumentId(state.annotationLayer)(ownProps.doc.id),
     scrollToSidebarComment: state.readerReducer.ui.pdf.scrollToSidebarComment,
     hidePdfSidebar: state.readerReducer.ui.pdf.hidePdfSidebar,
     error: state.readerReducer.ui.pdfSidebar.error,
